% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysCodesLis.R
\name{analysCodesList}
\alias{analysCodesList}
\title{analysCodesList}
\usage{
analysCodesList(dataS, codesLis)
}
\arguments{
\item{dataS}{a character}

\item{codesLis}{a character}
}
\value{
a list
}
\description{
analysCodesList
}
\examples{
codes=list("@essai@","@test@")
data = "this is an example @essai@, a bit long @essai@ text"
Result=analysCodesList(data,codes)
Result
}
