% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootStrap.R
\name{bootStrap}
\alias{bootStrap}
\title{bootStrap}
\usage{
bootStrap(nameDframe, grpSize)
}
\arguments{
\item{nameDframe}{a data.frame}

\item{grpSize}{a number}
}
\value{
a matrix
}
\description{
bootStrap
}
\examples{
j=c(10,14,56,30,58,78,99,1)
k=c(10,12,14,16,18,20,22,24)
x=data.frame(j,k)
res=bootStrap(x,5)
res
}
