% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxAndCox.R
\name{BoxAndCox}
\alias{BoxAndCox}
\title{BoxAndCox}
\usage{
BoxAndCox(rawVect, minLambda)
}
\arguments{
\item{rawVect}{a vector}

\item{minLambda}{a number}
}
\value{
a list
}
\description{
BoxAndCox
}
\examples{
vec=rlnorm(100, log(3), log(3))
BandC=BoxAndCox(vec, -3)
BandC
BAC=unlist(BandC$par)
BAC
rawVectBCFinal=unlist(subCalcBoxAndCox(vec, BandC$par))
}
