% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeCount.R
\name{codeCount}
\alias{codeCount}
\title{codeCount}
\usage{
codeCount(dataSet, code)
}
\arguments{
\item{dataSet}{a character}

\item{code}{a character}
}
\value{
a number
}
\description{
codeCount
}
\examples{
data = "this is an example @essai@"
codeCount(data, "@essai@") #number of lines containing the chain
}
