% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congruGen.R
\name{congruGen}
\alias{congruGen}
\title{congruGen}
\usage{
congruGen(seed, a)
}
\arguments{
\item{seed}{a number}

\item{a}{a number}
}
\value{
a list
}
\description{
congruGen
}
\examples{
seed = 123456789
for(i in 1: 10) {
Z=congruGen(seed)
seed=Z$seedUpdate
num=Z$aleaNum
print(num)
}

}
