% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooc.R
\name{cooc}
\alias{cooc}
\title{cooc}
\usage{
cooc(lines, code1, code2)
}
\arguments{
\item{lines}{character}

\item{code1}{character}

\item{code2}{character}
}
\value{
an integer
}
\description{
cooc
}
\examples{
lines ="Companies can boost responsiveness @performance@ by digital @digital@."
code1 = "@performance@"
code2 = "@digital@"
res=cooc(lines, code1, code2)
print(res)
}
