% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCodes.R
\name{loadCodes}
\alias{loadCodes}
\title{loadCodes}
\usage{
loadCodes(txtFile)
}
\arguments{
\item{txtFile}{a character}
}
\value{
a list
}
\description{
loadCodes
}
\examples{
theFile =system.file("codesList.txt", package = "codecountR")
data=loadCodes(theFile)
}
