% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medAbsDev.R
\name{medAbsDev}
\alias{medAbsDev}
\title{medAbsDev}
\usage{
medAbsDev(nameDframe)
}
\arguments{
\item{nameDframe}{a data.frame}
}
\value{
a data.frame
}
\description{
medAbsDev
}
\examples{
vec=rnorm(100, mean=10, sd=1)
vec2=rnorm(100, mean=15, sd=1)
df= data.frame(vec,vec2)
colnames(df) = c("vec", "vec2")
MAD=medAbsDev(df)
MAD
}
