% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piBlom.R
\name{piBlom}
\alias{piBlom}
\title{piBlom}
\usage{
piBlom(nameVect, sSize)
}
\arguments{
\item{nameVect}{a vector}

\item{sSize}{a number}
}
\value{
a vector
}
\description{
piBlom
}
\examples{
vec=c(1,2,3)
piBlom(vec, length(vec))
}
