% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piBloom.R
\name{piBloom}
\alias{piBloom}
\title{piBloom}
\usage{
piBloom(nameVect, sSize)
}
\arguments{
\item{nameVect}{a vector}

\item{sSize}{a number}
}
\value{
a vector
}
\description{
piBloom
}
\examples{
vec=c(1,2,3)
piBloom(vec, length(vec))
}
