% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subCalcBoxAndCox.R
\name{subCalcBoxAndCox}
\alias{subCalcBoxAndCox}
\title{subCalcBoxAndCox}
\usage{
subCalcBoxAndCox(sortedVect, actualLambda)
}
\arguments{
\item{sortedVect}{a vector}

\item{actualLambda}{a number}
}
\value{
a vector
}
\description{
subCalcBoxAndCox
}
\examples{
vec=rlnorm(100, log(3), log(3))
BandC=subCalcBoxAndCox(vec, -3)
}
