% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subCalcYeoAndJohnson.R
\name{subCalcYeoAndJohnson}
\alias{subCalcYeoAndJohnson}
\title{subCalcYeoAndJohnson}
\usage{
subCalcYeoAndJohnson(Vect, lambda)
}
\arguments{
\item{Vect}{a vector}

\item{lambda}{a number}
}
\value{
a vector
}
\description{
subCalcYeoAndJohnson
}
\examples{
vec=rlnorm(100, log(3), log(3))
subCalcYeoAndJohnson(vec, -3)
#only one lamba value tested
}
