% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPairs.R
\name{testPairs}
\alias{testPairs}
\title{testPairs}
\usage{
testPairs(listCodes, lines)
}
\arguments{
\item{listCodes}{character}

\item{lines}{character}
}
\value{
a list
}
\description{
testPairs
}
\examples{
#Co-occurrences computed line by line in the file. Structure the file accordingly.
#Multiple identical pairs on one line count as one unit.
lines =c("Companies can boost responsiveness @performance@ by digital @digital@.",
  "softwares @digital@ may reduce response time @performance@ improving @satisfaction@.")
listCodes=c("@satisfaction@", "@digital@", "@performance@")
coocurences = testPairs(listCodes, lines)
print(coocurences$matrix)
#save to file
#nameFile = paste("CooccurrenceMatrix_",format(Sys.time(),"\%d_\%m_\%Y-\%Hh\%Mm\%Ss"),".csv",sep = "")
#write.csv(coocurences$matrix, nameFile, row.names = TRUE)
}
