% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{tokenization}
\alias{tokenization}
\title{tokenization}
\usage{
tokenization(txtFile)
}
\arguments{
\item{txtFile}{a character}
}
\value{
a list
}
\description{
tokenization
}
\examples{
theFile =system.file("ExText.txt", package = "codecountR")
data=tokenization(theFile)
}
