% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify}
\alias{verify}
\title{verify}
\usage{
verify(lines, code1, code2)
}
\arguments{
\item{lines}{character}

\item{code1}{character}

\item{code2}{character}
}
\description{
verify
}
\examples{
lines ="Companies can boost responsiveness @performance@ by digital @digital@."
code1 = "@performance@"
code2 = "@digital@"
verify(lines,code1,code2)
}
