% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YeoAndJohnson.R
\name{YeoAndJohnson}
\alias{YeoAndJohnson}
\title{YeoAndJohnson}
\usage{
YeoAndJohnson(rawVect, minLambda)
}
\arguments{
\item{rawVect}{a vector}

\item{minLambda}{a number}
}
\value{
a list
}
\description{
YeoAndJohnson
}
\examples{
vec=rlnorm(100, log(3), log(3))
YandJ=YeoAndJohnson(vec, -3)
YandJ
YAJ=unlist(YandJ[3])
}
