% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zScore.R
\name{zScore}
\alias{zScore}
\title{zScore}
\usage{
zScore(nameDframe)
}
\arguments{
\item{nameDframe}{a data.frame}
}
\value{
a data.frame
}
\description{
zScore
}
\examples{
j=c(10,14,56,30,58,78,99,1)
k=c(10,12,14,16,18,20,22,24)
x=data.frame(j,k)
xZsc=zScore(x)
xZsc
}
